Check for unset variables.

+++ include/ldap_group_mapping.php
@@ -13,10 +13,10 @@ function map_ldap_groups($ldap, $info, $id) {
 	// Get group mappings
 	$mapping = $ldap -> config['group_mapping'];
 
-	// $key is group id. $lg is an array of 
+	// $key is group id. $lg is an array of
 	// LDAP groups that should map to $key.
 	foreach ($mapping as $key => $lg) {
-		
+
 		// Loop through each LDAP group
 		foreach ($lg as $i => $v) {
 
@@ -49,27 +49,31 @@ function map_ldap_groups($ldap, $info, $id) {
 	 *
 	 * This ensures that the user won't lose access
 	 * because they went from, for example, being an
-	 * undergrad to a graduate student. While their 
-	 * Piwigo group might not change, If we delete 
+	 * undergrad to a graduate student. While their
+	 * Piwigo group might not change, If we delete
 	 * last, they lose access to the Piwigo group.
 	 * Deleting first then updating prevents this.
 	 */
 
 	// The Delete
-	$query = '
-		DELETE FROM '. USER_GROUP_TABLE .'
-		WHERE user_id = ' . $uid . '
-		AND group_id IN('. implode(',', $removals) . ');';
+	if (isset($removals)) {
+		$query = '
+			DELETE FROM '. USER_GROUP_TABLE .'
+			WHERE user_id = ' . $uid . '
+			AND group_id IN('. implode(',', $removals) . ');';
 
-	pwg_query($query);
+		pwg_query($query);
+	}
 
 	// The Insert
-	mass_inserts(
-		USER_GROUP_TABLE,
-		array('group_id', 'user_id'),
-		$inserts,
-		array('ignore'=>true)
-	);
+	if (isset($inserts)) {
+		mass_inserts(
+			USER_GROUP_TABLE,
+			array('group_id', 'user_id'),
+			$inserts,
+			array('ignore'=>true)
+		);
+	}
 
 	// The Cleanup
 	include_once(PHPWG_ROOT_PATH.'admin/include/functions.php');
